Make library single-header
